requirements.txt
crewai
crewai_tools
langchain
langchain_community
langchain-openai
pyyaml
jinja2
python-dotenv
pytest